+++ modules/app/src/main/scalajs/LangoustineAppPlatform.scala
@@ -1,10 +1,8 @@
-package langoustine
-package lsp
-package fs2
+package langoustine.lsp.app
 
 import _root_.fs2 as FS2
 
-private[fs2] trait LangoustineAppPlatform:
+private[app] trait LangoustineAppPlatform:
   self: LangoustineApp.Config =>
 
   def in: FS2.Stream[cats.effect.IO, Byte] =
