+++ modules/app/src/main/scalanative/LangoustineAppPlatform.scala
@@ -1,14 +1,12 @@
-package langoustine
-package lsp
-package fs2
+package langoustine.lsp.app
 
 import _root_.fs2 as FS2
 import cats.effect.IO
 
-private[fs2] trait LangoustineAppPlatform:
+private[app] trait LangoustineAppPlatform:
   self: LangoustineApp.Config =>
 
   def inBufferSize: Int = 512
 
   def in: FS2.Stream[cats.effect.IO, Byte] =
-    FS2.io.stdin[IO](512)
+    FS2.io.stdin[IO](inBufferSize)
